requirements.txt
Flask>=2.1
Flask-Login>=0.6.2
Flask-Migrate>=4.0
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
Werkzeug>=2.1
mysqlclient>=2.1
itsdangerous>=2.1
transformers>=4.30.0
safetensors>=0.3.0
torch>=1.13.0
lime>=0.2.0
google-genai>=0.3.0
psutil>=5.9.0


